Fix cluster delete issue

+++ nacossync-worker/src/main/java/com/alibaba/nacossync/template/processor/ClusterDeleteProcessor.java
@@ -16,6 +16,9 @@
  */
 package com.alibaba.nacossync.template.processor;
 
+import com.alibaba.nacossync.dao.ClusterTaskAccessService;
+import com.alibaba.nacossync.dao.TaskAccessService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,22 +26,36 @@ import com.alibaba.nacossync.dao.ClusterAccessService;
 import com.alibaba.nacossync.pojo.result.ClusterDeleteResult;
 import com.alibaba.nacossync.pojo.request.ClusterDeleteRequest;
 import com.alibaba.nacossync.template.Processor;
+import org.springframework.util.CollectionUtils;
 
 /**
  * @author NacosSync
  * @version $Id: ClusterDeleteProcessor.java, v 0.1 2018-09-30 PM2:43 NacosSync Exp $$
  */
+@Slf4j
 @Service
 public class ClusterDeleteProcessor implements Processor<ClusterDeleteRequest, ClusterDeleteResult> {
 
     @Autowired
     private ClusterAccessService clusterAccessService;
 
+    @Autowired
+    private ClusterTaskAccessService clusterTaskAccessService;
+
+    @Autowired
+    private TaskAccessService taskAccessService;
+
     @Override
     public void process(ClusterDeleteRequest clusterDeleteRequest,
                         ClusterDeleteResult clusterDeleteResult, Object... others) throws Exception {
-
-        clusterAccessService.deleteByClusterId(clusterDeleteRequest.getClusterId());
-
+        if (!CollectionUtils.isEmpty(clusterTaskAccessService.findBySourceClusterId(clusterDeleteRequest.getClusterId())) ||
+                !CollectionUtils.isEmpty(clusterTaskAccessService.findByDestClusterId(clusterDeleteRequest.getClusterId())) ||
+                !CollectionUtils.isEmpty(taskAccessService.findBySourceClusterId(clusterDeleteRequest.getClusterId())) ||
+                !CollectionUtils.isEmpty(taskAccessService.findByDestClusterId(clusterDeleteRequest.getClusterId()))) {
+            log.error("该注册中心已配置同步任务，不可删除！");
+            throw new RuntimeException("该注册中心已配置同步任务，不可删除！");
+        } else {
+            clusterAccessService.deleteByClusterId(clusterDeleteRequest.getClusterId());
+        }
     }
 }
